Fix gesture issue and add markup

+++ StickerMator/StickerMatorModel.swift
@@ -15,16 +15,21 @@ struct StickerMatorModel: Codable {
     var stickers = [Sticker]()
     var mainImage: Data?  // main image to edit
     
-    // Sticker is an image
+    /// Sticker is actually an image
     struct Sticker: Identifiable, Hashable, Codable {
+        /// Image data saved as pngData()
         var data: Data
-        var x: Int
-        var y: Int
+        /// Abscissa position
+        var x: Float
+        /// Ordinate position
+        var y: Float
+        /// Sticker width
         var width: Int
+        /// Sticker height
         var height: Int
         let id: Int
         
-        fileprivate init( data: Data, x: Int, y: Int, width: Int, height: Int, id: Int) {
+        fileprivate init( data: Data, x: Float, y: Float, width: Int, height: Int, id: Int) {
             self.data = data
             self.x = x
             self.y = y
@@ -36,10 +41,12 @@ struct StickerMatorModel: Codable {
     
     init () {}
     
+    /// Initializes a new StickerMatorModel from json data.
     init(json: Data) throws {
         self = try JSONDecoder().decode(StickerMatorModel.self, from: json)
     }
     
+    /// Initializes a new StickerMatorModel from json data URL.
     init(url: URL) throws {
         let data = try Data(contentsOf: url)
         self = try StickerMatorModel(json: data)
@@ -47,12 +54,13 @@ struct StickerMatorModel: Codable {
     
     private var uniqueStickerId = 0
     
+    /// Encode self to JSON
     func jsonEncode() throws -> Data {
         return try JSONEncoder().encode(self)
     }
     
     
-    mutating func addSticker (imageData: Data, at location:(x: Int, y: Int), size: (width: Int, height: Int)) {
+    mutating func addSticker (imageData: Data, at location:(x: Float, y: Float), size: (width: Int, height: Int)) {
         uniqueStickerId += 1
         stickers.append(Sticker(data: imageData, x: location.x, y: location.y,
                                 width: size.width, height: size.height, id: uniqueStickerId))
